requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.0.6
django-mssql==1.8
djangorestframework==3.15.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
pyodbc==5.1.0
sqlparse==0.5.0
typing_extensions==4.11.0
Werkzeug==3.0.3
